diff components for tracking and preview
