+++ src/Pages/VideoPlayer/RenderWithTracking.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+const RenderWithTracking = ({ playlistID }) => {
+  return (
+    <div>
+      <h1>{playlistID}</h1>
+    </div>
+  );
+};
+
+export default RenderWithTracking;
